requirements.txt
django
celery
python-decouple
redis
